pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sophion"
version = "0.1.0"
description = "Personal knowledge engine with LLM-compiled wiki and challenger agent"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "markdownify>=0.14",
    "beautifulsoup4>=4.12",
    "python-frontmatter>=1.1",
    "rich>=13.0",
]

[project.scripts]
sophion = "sophion.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/sophion"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
